Suggested changes to main README.md
Apologies folks, I was a bit slow off the mark this morning reviewing the readme.

This is an alternative approach that I think might work a bit better. Basically, I don't think it should repeat/replace content from the curriculum introduction. After all, the point of this repo is to provide a preview of the curriculum. It is not part of the curriculum itself

I also thought it might be a good idea to clearly signpost the rationale for this piece of work, and give viewers some guidance as to the structure, given that there is no TOC page for them to use to navigate.

Also, the first heading was really a subheading; there wasn't really a proper intro and top-level  heading for the page. This also addresses that.